requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.21
Werkzeug==2.3.7
alembic==1.12.0
python-dotenv==1.0.0
pytz==2023.3.post1
email-validator==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
APScheduler==3.10.4